+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * [FEAT] Support `sensitive: true` on subfields (with `on:`)
 * [FEAT] Support `preprocess` on subfields (with `on:`)
 * [FEAT] Support `default` on subfields (with `on:`)
+* [FEAT] Added `#done!` method for early completion with success result
 
 ## 0.1.0-alpha.2.8.1
 * [BUGFIX] Fixed symbol callback and message handlers not working in inherited classes due to private method visibility issues